Updating example.js
